discordbot/requirements.txt
discord
pyyaml
aiohttp
pillow
matplotlib